Consolidate syncAndWait into shared common.ts helper
Replace 11 duplicate syncAndWait functions across helper files with a
single shared implementation in common.ts. The unified function includes
a stability re-check (waits 1s and re-verifies "Status: Success") to
guard against false positives from transient sync states. Accepts
optional healthCenter and timeout parameters.

[ci skip]
