+++ client/e2e/acute-illness-encounter-chw.spec.ts
@@ -6,6 +6,7 @@ import {
 } from './helpers/case-management';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   createAdultAndStartEncounter,
   completeSymptoms,
@@ -14,7 +15,6 @@ import {
   completeLaboratory,
   completeNextSteps,
   endEncounter,
-  syncAndWait,
   queryAcuteIllnessNodes,
   backdateAcuteIllnessEncounter,
   navigateToParticipantPage,
