+++ client/e2e/child-scoreboard-encounter-chw.spec.ts
@@ -2,12 +2,12 @@ import { test, expect } from '@playwright/test';
 import { setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   createChildAndStartEncounter,
   completeNCDA,
   completeVaccinationHistory,
   endChildScoreboardEncounter,
-  syncAndWait,
   queryChildScoreboardNodes,
 } from './helpers/child-scoreboard';
