+++ client/e2e/education-session-chw.spec.ts
@@ -2,13 +2,13 @@ import { test, expect } from '@playwright/test';
 import { setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   navigateToEducationSession,
   selectTopics,
   toggleAllParticipants,
   selectParticipant,
   endEducationSession,
-  syncAndWait,
   queryEducationSessionNodes,
 } from './helpers/education-session';
