+++ client/e2e/group-session-chw.spec.ts
@@ -2,6 +2,7 @@ import { test, expect } from '@playwright/test';
 import { click, setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   navigateToChwGroupSession,
   createMotherOnAttendancePage,
@@ -18,7 +19,6 @@ import {
   completeNutritionSigns,
   completeFamilyPlanning,
   endGroupSession,
-  syncAndWait,
   queryGroupSessionNodes,
 } from './helpers/group-session';
