+++ client/e2e/group-session-nurse.spec.ts
@@ -2,6 +2,7 @@ import { test, expect } from '@playwright/test';
 import { click, setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   navigateToNurseGroupSession,
   createMotherOnAttendancePage,
@@ -26,7 +27,6 @@ import {
   completeLactation,
   completeMotherFbf,
   endGroupSession,
-  syncAndWait,
   queryGroupSessionNodes,
 } from './helpers/group-session';
