+++ client/e2e/helpers/ncd.ts
@@ -1106,37 +1106,6 @@ export async function leaveRecurrentEncounter(page: Page) {
 }
 
 // ---------------------------------------------------------------------------
-// Sync (reuse from nutrition.ts pattern)
-// ---------------------------------------------------------------------------
-
-/**
- * Sync data and wait for success.
- * Click sync icon → device status → wait for success → go back.
- */
-export async function syncAndWait(page: Page) {
-  // Navigate to device status.
-  await click(page.locator('span.sync-icon'), page);
-
-  // Wait for the Device Status page to be rendered.
-  await page.locator('.device-status').waitFor({ timeout: 10000 });
-
-  // Find the health center section for "Nyange Health Center".
-  const hcSection = page.locator('.health-center', {
-    has: page.locator('h2', { hasText: 'Nyange Health Center' }),
-  });
-  await hcSection.waitFor({ timeout: 10000 });
-
-  // Wait for sync success.
-  await hcSection
-    .locator('.sync-status', { hasText: 'Status: Success' })
-    .waitFor({ timeout: 120000 });
-
-  // Go back.
-  await page.goBack();
-  await page.waitForTimeout(1000);
-}
-
-// ---------------------------------------------------------------------------
 // Backend verification via drush
 // ---------------------------------------------------------------------------
