+++ client/e2e/helpers/well-child.ts
@@ -994,27 +994,6 @@ export async function endWellChildEncounter(page: Page) {
 }
 
 // ---------------------------------------------------------------------------
-// Sync helper
-// ---------------------------------------------------------------------------
-
-export async function syncAndWait(page: Page) {
-  await click(page.locator('span.sync-icon'), page);
-
-  await page.locator('.device-status').waitFor({ timeout: 10000 });
-
-  const nyange = page.locator('.health-center', {
-    has: page.locator('h2', { hasText: 'Nyange Health Center' }),
-  });
-  await nyange.waitFor({ timeout: 10000 });
-
-  await nyange
-    .locator('.sync-status', { hasText: 'Status: Success' })
-    .waitFor({ timeout: 480000 });
-
-  await page.goBack();
-}
-
-// ---------------------------------------------------------------------------
 // Backend verification via drush
 // ---------------------------------------------------------------------------
