+++ client/e2e/nutrition-encounter-chw.spec.ts
@@ -7,6 +7,7 @@ import {
 } from './helpers/case-management';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   createChildAndStartEncounter,
   enterWeight,
@@ -14,7 +15,6 @@ import {
   enterNutritionSigns,
   saveActivity,
   endEncounter,
-  syncAndWait,
   completeSendToHC,
   completeHealthEducation,
   completeContributingFactors,
