+++ client/e2e/nutrition-encounter.spec.ts
@@ -2,6 +2,7 @@ import { test, expect } from '@playwright/test';
 import { click, setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   createChildAndStartEncounter,
   completeNCDA,
@@ -11,7 +12,6 @@ import {
   enterNutritionSigns,
   saveActivity,
   endEncounter,
-  syncAndWait,
   completeSendToHC,
   completeHealthEducation,
   completeContributingFactors,
