+++ client/e2e/prenatal-encounter-chw.spec.ts
@@ -6,7 +6,7 @@ import {
 } from './helpers/case-management';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
-import { syncAndWait } from './helpers/nutrition';
+import { syncAndWait } from './helpers/common';
 import {
   createAdultFemaleAndStartEncounter,
   startPrenatalEncounter,
