+++ client/e2e/stock-management-nurse.spec.ts
@@ -2,11 +2,11 @@ import { test, expect } from '@playwright/test';
 import { setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   navigateToStockManagement,
   completeReceiveStock,
   completeCorrectEntry,
-  syncAndWait,
   queryStockUpdateNodes,
 } from './helpers/stock-management';
 import { click } from './helpers/auth';
