+++ client/e2e/well-child-encounter-chw.spec.ts
@@ -2,6 +2,7 @@ import { test, expect } from '@playwright/test';
 import { setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   createChildAndStartWellChildEncounter,
   completePregnancySummary,
@@ -11,7 +12,6 @@ import {
   completeHomeVisit,
   completeNextSteps,
   endWellChildEncounter,
-  syncAndWait,
   queryWellChildNodes,
 } from './helpers/well-child';
