+++ client/e2e/well-child-encounter-nurse.spec.ts
@@ -2,6 +2,7 @@ import { test, expect } from '@playwright/test';
 import { setupDevice } from './helpers/auth';
 import { installCursorScript } from './helpers/cursor';
 import { resetDevice } from './helpers/device';
+import { syncAndWait } from './helpers/common';
 import {
   createChildAndStartWellChildEncounter,
   completeDangerSigns,
@@ -12,7 +13,6 @@ import {
   completeNCDA,
   completeNextSteps,
   endWellChildEncounter,
-  syncAndWait,
   queryWellChildNodes,
 } from './helpers/well-child';
